Refactored to support polling both indexd and bucket
